chore(docs): Optimized array indexing in the command bindings example
# Project State

## Current Focus
Optimized array indexing in the command bindings example

## Context
The change improves performance by removing unnecessary type casting in array index calculations during plane blitting operations.

## Completed
- [x] Removed redundant `as usize` cast in destination index calculation
- [x] Simplified index calculation while maintaining safety checks

## In Progress
- [x] Performance optimization for plane blitting operations

## Blockers
- None identified

## Next Steps
1. Verify no regression in visual output
2. Check for additional performance optimizations in related code
---
category: chore
scope: docs
---

+++ examples/_cookbook/command_bindings.rs
@@ -164,7 +164,7 @@ fn blit_plane(src: &Plane, dst: &mut Plane, dst_x: u16, dst_y: u16) {
             if dx >= dst.width || dy >= dst.height {
                 continue;
             }
-            let dst_idx = (dy as usize * dst.width as usize + dx as usize) as usize;
+            let dst_idx = dy as usize * dst.width as usize + dx as usize;
             if dst_idx < dst.cells.len() {
                 dst.cells[dst_idx] = src.cells[src_idx].clone();
             }
